style and movement changes
