+++ src/pages/groups/List.js
@@ -2,7 +2,7 @@ import _ from 'lodash/fp'
 import { Fragment } from 'react'
 import { a, b, div, h } from 'react-hyperscript-helpers'
 import { pure } from 'recompose'
-import { buttonPrimary, Clickable, linkButton, search, spinnerOverlay } from 'src/components/common'
+import { buttonPrimary, Clickable, linkButton, search, spinnerOverlay, LargeFadeBox } from 'src/components/common'
 import { icon } from 'src/components/icons'
 import { validatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
@@ -17,7 +17,6 @@ import * as Utils from 'src/libs/utils'
 import { Component } from 'src/libs/wrapped-components'
 import { styles } from 'src/pages/groups/common'
 import { validate } from 'validate.js'
-import { LargeFadeBox } from 'src/pages/workspaces/List'
 
 
 const groupNameValidator = existing => ({
